requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.2
djangorestframework==3.12.4
drf-nested-routers==0.93.3
flake8==3.9.1
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
